+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/CascadeFilters/CascadeFilterControl.tsx
@@ -19,8 +19,8 @@
 import React from 'react';
 import { styled, DataMask } from '@superset-ui/core';
 import Icon from 'src/components/Icon';
-import FilterControl from './FilterControl';
-import { Filter } from '../types';
+import FilterControl from '../FilterControl/FilterControl';
+import { Filter } from '../../types';
 import { CascadeFilter } from './types';
 
 interface CascadeFilterControlProps {
