+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/CascadeFilters/CascadePopover.tsx
@@ -24,10 +24,10 @@ import { Pill } from 'src/dashboard/components/FiltersBadge/Styles';
 import { useSelector } from 'react-redux';
 import { getInitialMask } from 'src/dataMask/reducer';
 import { MaskWithId } from 'src/dataMask/types';
-import FilterControl from './FilterControl';
+import FilterControl from '../FilterControl/FilterControl';
 import CascadeFilterControl from './CascadeFilterControl';
 import { CascadeFilter } from './types';
-import { Filter } from '../types';
+import { Filter } from '../../types';
 
 interface CascadePopoverProps {
   filter: CascadeFilter;
