+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/utils.ts
@@ -18,7 +18,6 @@
  */
 
 import { Filter } from '../types';
-import { CascadeFilter } from './types';
 
 export function mapParentFiltersToChildren(
   filters: Filter[],
@@ -34,20 +33,4 @@ export function mapParentFiltersToChildren(
     }
   });
   return cascadeChildren;
-}
-
-export function buildCascadeFiltersTree(filters: Filter[]): CascadeFilter[] {
-  const cascadeChildren = mapParentFiltersToChildren(filters);
-
-  const getCascadeFilter = (filter: Filter): CascadeFilter => {
-    const children = cascadeChildren[filter.id] || [];
-    return {
-      ...filter,
-      cascadeChildren: children.map(getCascadeFilter),
-    };
-  };
-
-  return filters
-    .filter(filter => !filter.cascadeParentIds?.length)
-    .map(getCascadeFilter);
 }
